Add 'oracle-java' role dependency as git submodule
The module tracks https://github.com/cwardgar/oracle-java, branch 'mine'. It was forked from https://github.com/ansiblebit/oracle-java, branch 'master'.

Starting with 'oracle-java', we will be adding our role dependencies directly to the project as git submodules, instead of downloading them from
Ansible Galaxy. We're doing this so that we can make changes to those roles, if necessary. As a result, we don't need requirements.yml any more.

Also, we don't need to maintain 'dependencies' in nexus/meta/main.yml. Arguably, we never needed to maintain it in the first place, because we
don't plan to publish our role to Galaxy. Furthermore, Ansible isn't smart enough to auto-download the roles defined there.