Added Ansible configuration file to /provisioning/
Suppress display of "include" and "skipped" tasks. Basically, omit all cyan text.
This work was based on third-party code. I've documented the license of the original work
and the changes that I made.

Suppress generation of *.retry files.

Update test script so that ansible-playbook will be able to find ansible.cfg when it is
run inside the Docker container. Also refactored the "docker run" command to include
more useful comments.

Update to the latest commit of cwardgar/ansible-nexus3-oss.