Move all testing content into dedicated Ansible role.
Create new testing playbook and run all tests from there. The main playbook no longer includes any test tasks.

Make an explicit inventory (including variables) for Docker hosts. Previously, I was relying on inventory config done in the docker-ubuntu1604-ansible image.

Update documentation to account for the reorganization.

In test.sh:
* Mount the provisioning directory, not the project directory. That simplified things somewhat.
* Use the new Docker inventory file, instead of relying on the /etc/ansible/hosts file setup by the image.
* Account for different location of the script.
* Run the test playbook at the end.

Add some 'ansible-nexus3-oss' variables to /provisioning/vars/main.yml, not because we needed to override the values, but because we needed to make the variables visible to the 'testing' role.

Updated to latest commit of cwardgar/ansible-nexus3-oss.